Change `dims` type from `uint64_t` to `int64_t` throughout the codebase for consistency, update related methods, diagnostics, and tests for negative/large index handling.

+++ src/Sema/Ast/Sema.Literal.cpp
@@ -518,7 +518,7 @@ Result AstArrayLiteral::semaPostNode(Sema& sema)
             values.push_back(sema.constantRefOf(nodeView.nodeRef));
     }
 
-    const std::vector dims         = {elements.size()};
+    const std::vector dims         = {static_cast<int64_t>(elements.size())};
     const TypeRef     arrayTypeRef = sema.typeMgr().addType(TypeInfo::makeArray(dims, refElemTypeRef));
     sema.setType(sema.curNodeRef(), arrayTypeRef);
